refactor(napari_curvealign): split UI into widgets/ subpackage (#48)
- widgets/widget_utils.py: luma + ROI save format helpers
- widgets/dialogs.py: BoundaryType, advanced/results/metrics dialogs
- widgets/curve_align_widget.py: CurveAlignWidget + create_curve_align_widget
- Remove napari_curvealign/widget.py; plugin and examples import from widgets

Behavior matches pre-split monolith (mechanical move + helper renames only).
Layout inspired by napari-imagej widgets package.

Made-with: Cursor
